+++ cmd/frontend/internal/pkg/useractivity/useractivity.go
@@ -7,6 +7,8 @@ package useractivity
 
 import (
 	"context"
+	"fmt"
+	"math"
 	"math/rand"
 	"os"
 	"strconv"
@@ -25,12 +27,20 @@ var (
 
 	keyPrefix = "user_activity:"
 	pool      *redis.Pool
+
+	timeNow = time.Now
 )
 
 const (
-	fPageViews     = "pageviews"
-	fLastActive    = "lastactive"
-	fSearchQueries = "searchqueries"
+	fPageViews                     = "pageviews"
+	fLastActive                    = "lastactive"
+	fSearchQueries                 = "searchqueries"
+	fCodeIntelActions              = "codeintelactions"
+	fLastActiveCodeHostIntegration = "lastactivecodehostintegration"
+
+	daysOfHistory   = 14
+	weeksOfHistory  = 10
+	monthsOfHistory = 3
 )
 
 func init() {
@@ -117,17 +127,17 @@ func GetByUserID(userID int32) (*types.UserActivity, error) {
 	key := keyPrefix + userIDStr
 
 	c := pool.Get()
-	values, err := redis.Values(c.Do("HMGET", key, fPageViews, fSearchQueries, fLastActive))
+	values, err := redis.Values(c.Do("HMGET", key, fPageViews, fSearchQueries, fLastActive, fCodeIntelActions, fLastActiveCodeHostIntegration))
 	c.Close()
 	if err != nil && err != redis.ErrNil {
 		return nil, err
 	}
 
-	var lastActiveStr string
+	var lastActiveStr, lastActiveCodeHostStr string
 	a := &types.UserActivity{
 		UserID: userID,
 	}
-	_, err = redis.Scan(values, &a.PageViews, &a.SearchQueries, &lastActiveStr)
+	_, err = redis.Scan(values, &a.PageViews, &a.SearchQueries, &lastActiveStr, &a.CodeIntelligenceActions, &lastActiveCodeHostStr)
 	if err != nil && err != redis.ErrNil {
 		return nil, err
 	}
@@ -137,53 +147,181 @@ func GetByUserID(userID int32) (*types.UserActivity, error) {
 		if err != nil {
 			return nil, err
 		}
-		a.LastPageViewTime = &t
+		a.LastActiveTime = &t
+	}
+
+	if lastActiveCodeHostStr != "" {
+		t, err := time.Parse(time.RFC3339, lastActiveCodeHostStr)
+		if err != nil {
+			return nil, err
+		}
+		a.LastCodeHostIntegrationTime = &t
 	}
 
 	return a, nil
 }
 
-// GetUsersActiveTodayCount returns a count of users that have had pageviews
-// so far today.
+type SiteActivityOptions struct {
+	Days   int
+	Weeks  int
+	Months int
+}
+
+func minIntOrZero(a, b int) int {
+	min := b
+	if a < b {
+		min = a
+	}
+	if min < 0 {
+		return 0
+	}
+	return min
+}
+
+// GetSiteActivity returns the current site's SiteActivity
+func GetSiteActivity(opt *SiteActivityOptions) (*types.SiteActivity, error) {
+	var (
+		days   = daysOfHistory
+		weeks  = weeksOfHistory
+		months = monthsOfHistory
+	)
+
+	if opt != nil {
+		days = minIntOrZero(opt.Days, daysOfHistory)
+		weeks = minIntOrZero(opt.Weeks, weeksOfHistory)
+		months = minIntOrZero(opt.Months, monthsOfHistory)
+	}
+
+	daus, err := DAUs(days)
+	if err != nil {
+		return nil, err
+	}
+	waus, err := WAUs(weeks)
+	if err != nil {
+		return nil, err
+	}
+	maus, err := MAUs(months)
+	if err != nil {
+		return nil, err
+	}
+	return &types.SiteActivity{
+		DAUs: daus,
+		WAUs: waus,
+		MAUs: maus,
+	}, nil
+}
+
+// GetUsersActiveTodayCount returns a count of users that have been active today.
 func GetUsersActiveTodayCount() (int, error) {
 	c := pool.Get()
 	defer c.Close()
 
-	count, err := redis.Int(c.Do("SCARD", usersActiveKey()))
+	count, err := redis.Int(c.Do("SCARD", usersActiveKeyFromDaysAgo(0)))
 	if err == redis.ErrNil {
 		err = nil
 	}
 	return count, err
 }
 
-// LogPageView increments a user's pageview count.
-func LogPageView(isAuthenticated bool, userID int32, userCookieID string) error {
+// calcUniques calculates the number of unique users starting at 00:00:00 on a given UTC date over a
+// period of time (years, months, and days).
+func calcUniques(c redis.Conn, dayStart time.Time, years int, months int, days int) (*types.SiteActivityPeriod, error) {
+	allUniqueUserIds := map[string]bool{}
+	registeredUserIds := map[string]bool{}
+
+	dayStart = time.Date(dayStart.Year(), dayStart.Month(), dayStart.Day(), 0, 0, 0, 0, time.UTC)
+	d := dayStart.AddDate(years, months, days).AddDate(0, 0, -1)
+
+	// Start at the end date, and loop backwards until reaching the start
+	for d.After(dayStart) || d.Equal(dayStart) {
+		values, err := redis.Values(c.Do("SMEMBERS", usersActiveKeyFromDate(d)))
+		if err != nil && err != redis.ErrNil {
+			return nil, err
+		}
+		for _, id := range values {
+			bid := id.([]byte)
+			sid := string(bid)
+			allUniqueUserIds[sid] = true
+			if len(bid) != 36 { // id is a numerical Sourcegraph user id, not an anonymous user's Telligent cookie id
+				registeredUserIds[sid] = true
+			}
+		}
+
+		d = d.AddDate(0, 0, -1)
+	}
+
+	return &types.SiteActivityPeriod{
+		StartTime:           dayStart,
+		UserCount:           int32(len(allUniqueUserIds)),
+		RegisteredUserCount: int32(len(registeredUserIds)),
+		AnonymousUserCount:  int32(len(allUniqueUserIds)) - int32(len(registeredUserIds)),
+	}, nil
+}
+
+// DAUs returns a count of daily active users for the last daysCount days (including the current, partial day).
+func DAUs(daysCount int) ([]*types.SiteActivityPeriod, error) {
 	c := pool.Get()
 	defer c.Close()
 
-	uniqueID := userCookieID
-
-	// If the user is authenticated, store their activity in the appropriate user ID-keyed cache.
-	if isAuthenticated {
-		userIDStr := strconv.Itoa(int(userID))
-		uniqueID = userIDStr
-		key := keyPrefix + uniqueID
-
-		// Increment the user's pageview count.
-		if err := c.Send("HINCRBY", key, fPageViews, 1); err != nil {
-			return err
+	var daus []*types.SiteActivityPeriod
+	now := timeNow().UTC()
+	for daysAgo := 0; daysAgo < daysCount; daysAgo++ {
+		uniques, err := calcUniques(c, now.AddDate(0, 0, -daysAgo), 0, 0, 1)
+		if err != nil {
+			return nil, err
 		}
-		if err := c.Send("HSET", key, fLastActive, time.Now().Format(time.RFC3339)); err != nil {
-			return err
-		}
+		daus = append(daus, uniques)
 	}
-
-	// Regardless of authenicatation status, add the user's unique ID to the set of active users.
-	return c.Send("SADD", usersActiveKey(), uniqueID)
+	return daus, nil
 }
 
-// LogSearchQuery increments a user's search query count.
-func LogSearchQuery(userID int32) error {
+// WAUs returns a count of daily active users for the last weeksCount calendar weeks (including the current, partial week).
+func WAUs(weeksCount int) ([]*types.SiteActivityPeriod, error) {
+	c := pool.Get()
+	defer c.Close()
+
+	var waus []*types.SiteActivityPeriod
+
+	for w := 0; w < weeksCount; w++ {
+		weekStartDate := startOfWeek(w)
+		uniques, err := calcUniques(c, weekStartDate, 0, 0, 7)
+		if err != nil {
+			return nil, err
+		}
+		waus = append(waus, uniques)
+	}
+	return waus, nil
+}
+
+// MAUs returns a count of daily active users for the last monthsCount calendar months (including the current, partial month).
+func MAUs(monthsCount int) ([]*types.SiteActivityPeriod, error) {
+	c := pool.Get()
+	defer c.Close()
+
+	var maus []*types.SiteActivityPeriod
+
+	for m := 0; m < monthsCount; m++ {
+		monthStartDate := startOfMonth(m)
+		uniques, err := calcUniques(c, monthStartDate, 0, 1, 0)
+		if err != nil {
+			return nil, err
+		}
+		maus = append(maus, uniques)
+	}
+	return maus, nil
+}
+
+// logPageView increments a user's pageview count.
+func logPageView(userID int32) error {
+	key := keyPrefix + strconv.Itoa(int(userID))
+	c := pool.Get()
+	defer c.Close()
+
+	return c.Send("HINCRBY", key, fPageViews, 1)
+}
+
+// logSearchQuery increments a user's search query count.
+func logSearchQuery(userID int32) error {
 	key := keyPrefix + strconv.Itoa(int(userID))
 	c := pool.Get()
 	defer c.Close()
@@ -191,18 +329,108 @@ func LogSearchQuery(userID int32) error {
 	return c.Send("HINCRBY", key, fSearchQueries, 1)
 }
 
-func usersActiveKey() string {
-	return keyPrefix + ":usersactive:" + time.Now().Format("2006-01-02")
+// logCodeIntel increments a user's code intelligence usage count.
+func logCodeIntelAction(userID int32) error {
+	key := keyPrefix + strconv.Itoa(int(userID))
+	c := pool.Get()
+	defer c.Close()
+
+	return c.Send("HINCRBY", key, fCodeIntelActions, 1)
 }
 
-// gc expires active user sets after a week. We only use the current day, but
-// a weeks worth of data may be useful in the future for analysis.
+// logCodeHostIntegrationUsage logs the last time a user was active on a code host integration
+func logCodeHostIntegrationUsage(userID int32) error {
+	key := keyPrefix + strconv.Itoa(int(userID))
+	c := pool.Get()
+	defer c.Close()
+
+	now := timeNow().UTC()
+	return c.Send("HSET", key, fLastActiveCodeHostIntegration, now.Format(time.RFC3339))
+}
+
+// LogActivity logs any user activity (page view, integration usage, etc) to their "last active" time, and
+// adds their unique ID to the set of active users
+func LogActivity(isAuthenticated bool, userID int32, userCookieID string, event string) error {
+	c := pool.Get()
+	defer c.Close()
+
+	uniqueID := userCookieID
+
+	// If the user is authenticated, set uniqueID to their user ID, and store their "last active time" in the
+	// appropriate user ID-keyed cache.
+	if isAuthenticated {
+		userIDStr := strconv.Itoa(int(userID))
+		uniqueID = userIDStr
+		key := keyPrefix + uniqueID
+
+		// Set the user's last active time
+		now := timeNow().UTC()
+		if err := c.Send("HSET", key, fLastActive, now.Format(time.RFC3339)); err != nil {
+			return err
+		}
+	}
+
+	// Regardless of authenicatation status, add the user's unique ID to the set of active users.
+	if err := c.Send("SADD", usersActiveKeyFromDaysAgo(0), uniqueID); err != nil {
+		return err
+	}
+
+	// If the user isn't authenticated, return at this point and don't record user-level properties.
+	if !isAuthenticated {
+		return nil
+	}
+
+	switch event {
+	case "SEARCHQUERY":
+		return logSearchQuery(userID)
+	case "PAGEVIEW":
+		return logPageView(userID)
+	case "CODEINTEL":
+		return logCodeIntelAction(userID)
+	case "CODEINTELINTEGRATION":
+		if err := logCodeHostIntegrationUsage(userID); err != nil {
+			return err
+		}
+		return logCodeIntelAction(userID)
+	}
+	return fmt.Errorf("unknown user event %s", event)
+}
+
+func usersActiveKeyFromDate(date time.Time) string {
+	return keyPrefix + ":usersactive:" + time.Date(date.Year(), date.Month(), date.Day(), 0, 0, 0, 0, time.UTC).Format("2006-01-02")
+}
+
+func usersActiveKeyFromDaysAgo(daysAgo int) string {
+	now := timeNow().UTC()
+	return usersActiveKeyFromDate(now.AddDate(0, 0, -daysAgo))
+}
+
+func startOfWeek(weeksAgo int) time.Time {
+	if weeksAgo > 0 {
+		return startOfWeek(0).AddDate(0, 0, -7*weeksAgo)
+	}
+
+	// If weeksAgo == 0, start at timeNow(), and loop back by day until we hit a Sunday
+	now := timeNow().UTC()
+	date := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, time.UTC)
+	for date.Weekday() != time.Sunday {
+		date = date.AddDate(0, 0, -1)
+	}
+	return date
+}
+
+func startOfMonth(monthsAgo int) time.Time {
+	now := timeNow().UTC()
+	return time.Date(now.Year(), now.Month(), 1, 0, 0, 0, 0, time.UTC).AddDate(0, -monthsAgo, 0)
+}
+
+// gc expires active user sets after the max of daysOfHistory, weeksOfHistory, and monthsOfHistory have passed.
 func gc() {
 	for {
-		key := usersActiveKey()
+		key := usersActiveKeyFromDaysAgo(0)
 
 		c := pool.Get()
-		err := c.Send("EXPIRE", key, 60*60*24*7) // 1 week
+		err := c.Send("EXPIRE", key, 60*60*24*int(math.Max(31*monthsOfHistory, math.Max(7*weeksOfHistory, daysOfHistory))))
 		c.Close()
 
 		if err != nil {