+++ web/src/auth/SignInPage.tsx
@@ -138,7 +138,7 @@ interface SignInPageProps {
 
 export class SignInPage extends React.Component<SignInPageProps> {
     public componentDidMount(): void {
-        eventLogger.logViewEvent('SignIn')
+        eventLogger.logViewEvent('SignIn', {}, false)
     }
 
     public render(): JSX.Element | null {